아이스크림 만들기 함수를 새로운 파일에 분리하여 저장
