+++ CodeStarterCamp_Week1/drawIceCream.swift
@@ -1,0 +1,20 @@
+//
+//  iceCream.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by 이주영 on 2022/04/22.
+//
+
+import Foundation
+
+func makeIceCream (height: Int) {
+    for _ in 1...height {
+        print("***********")
+    }
+}
+
+func makeBar (height: Int) {
+    for _ in 1...height {
+        print("    | |")
+    }
+}
