+++ CodeStarterCamp_Week1/iceCream.swift
@@ -1,0 +1,8 @@
+//
+//  iceCream.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by 이주영 on 2022/04/22.
+//
+
+import Foundation
