+++ CodeStarterCamp_Week1/main.swift
@@ -8,18 +8,6 @@
 
 import Foundation
 
-func makeIceCream (height: Int) {
-    for _ in 1...height {
-        print("***********")
-    }
-}
-
-func makeBar (height: Int) {
-    for _ in 1...height {
-        print("    | |")
-    }
-}
-
 makeIceCream(height: 8)
 makeBar(height: 4)
